Add consumerProguardFiles for lib. Clear item tint list if user no specify one.
