+++ app/build.gradle
@@ -32,8 +32,8 @@ dependencies {
     implementation "com.android.support:appcompat-v7:$rootProject.supportLibraryVersion"
     implementation "com.android.support:design:$rootProject.supportLibraryVersion"
 
-//    implementation project(':widget')
-    implementation 'com.github.ittianyu:BottomNavigationViewEx:2.0.2'
+    implementation project(':widget')
+//    implementation 'com.github.ittianyu:BottomNavigationViewEx:2.0.2'
     // leakcanary
     debugImplementation 'com.squareup.leakcanary:leakcanary-android:1.5.4'
     releaseImplementation 'com.squareup.leakcanary:leakcanary-android-no-op:1.5.4'
