+++ app/src/main/res/drawable/ic_face_unhappy.xml
@@ -4,6 +4,6 @@
     android:viewportHeight="1024.0"
     android:viewportWidth="1024.0">
     <path
-        android:fillColor="#000000"
+        android:fillColor="#ccc"
         android:pathData="M512,0a512,512 0,1 0,512 512A512.8,512.8 0,0 0,512 0zM512,956.4A444.4,444.4 0,1 1,956.4 512,444.7 444.7,0 0,1 512,956.4zM531.9,613.9a233.5,233.5 0,0 0,-74.1 12c-33.6,16.8 -87.1,42.3 -122,94.5 -10.9,16.2 -15,25.4 -12.2,34.6 4.9,16.6 30.7,28.9 42.9,22.3a11.7,11.7 0,0 0,3.5 -2.9c34.8,-53.2 86.9,-99.9 150.2,-99.9 32.5,0 68.3,4.5 94.1,22.4a416.8,416.8 0,0 1,52.1 47.8l28.5,29.5a75.1,75.1 0,0 0,31.2 -12.8,27.3 27.3,0 0,0 12.4,-13.7c0.5,-1.9 2.9,-11.4 -10.3,-27.3 -87.9,-106.4 -196.3,-106.5 -196.3,-106.5zM705.5,344.5a65,65 0,1 0,65.1 65.1,65.1 65.1,0 0,0 -65.1,-65.1zM428,409.6a65,65 0,1 0,-65.1 65A65.1,65.1 0,0 0,428 409.6z" />
 </vector>
