+++ app/src/main/res/drawable/ic_lamp_close.xml
@@ -1,4 +1,4 @@
 <vector android:height="24dp" android:viewportHeight="1024.0"
     android:viewportWidth="1024.0" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#404040" android:pathData="M515.3,65.1h-6.6c-162.9,0.2 -308.3,131.1 -308.8,295.8 0,81.9 34.7,146.2 69.6,194.6 35,47.6 70.8,84.9 73.3,99 6,21.6 5.9,37.4 9.1,53.7 2.1,16.4 15.8,35.2 37.6,40.9 23.1,7 71.4,9 122.5,9.3 51.1,-0.3 99.4,-2.3 122.5,-9.3 21.9,-5.6 35.6,-24.4 37.7,-40.9 3.2,-16.2 3,-32 9,-53.6 2.5,-14.2 38.4,-51.5 73.4,-99.1 35,-48.5 69.6,-112.7 69.5,-194.6C823.7,196.2 678.2,65.3 515.3,65.1zM725.8,534.8c-32.8,46.6 -68.1,77 -78.7,110.3 -7.1,25.8 -7.4,45.6 -9.6,55.9 -3.4,10.2 -0.7,8.7 -13.8,14.3 -11.9,4.3 -63.2,7.9 -111.7,7.6 -48.5,0.3 -99.8,-3.3 -111.6,-7.6 -13,-5.6 -10.3,-4.2 -13.8,-14.3 -2.3,-10.3 -2.5,-30.1 -9.6,-55.9 -10.6,-33.2 -45.9,-63.7 -78.7,-110.3 -32.9,-45.8 -62.9,-102.3 -62.9,-173.9 0,-71.4 31.9,-136.3 82.3,-183.7 50.4,-47.4 119,-76.6 190.9,-76.6h6.6c72,0 140.7,29.2 191,76.6 50.4,47.4 82.3,112.2 82.3,183.7 0,71.5 -29.9,128.1 -62.9,173.9zM381.2,824.7h261.7v-35.5L381.2,789.2v35.5zM381.2,891.8h261.7v-35.6L381.2,856.2v35.6zM454.6,958.9L569.5,958.9v-35.6L454.6,923.3v35.6z"/>
+    <path android:fillColor="#ccc" android:pathData="M515.3,65.1h-6.6c-162.9,0.2 -308.3,131.1 -308.8,295.8 0,81.9 34.7,146.2 69.6,194.6 35,47.6 70.8,84.9 73.3,99 6,21.6 5.9,37.4 9.1,53.7 2.1,16.4 15.8,35.2 37.6,40.9 23.1,7 71.4,9 122.5,9.3 51.1,-0.3 99.4,-2.3 122.5,-9.3 21.9,-5.6 35.6,-24.4 37.7,-40.9 3.2,-16.2 3,-32 9,-53.6 2.5,-14.2 38.4,-51.5 73.4,-99.1 35,-48.5 69.6,-112.7 69.5,-194.6C823.7,196.2 678.2,65.3 515.3,65.1zM725.8,534.8c-32.8,46.6 -68.1,77 -78.7,110.3 -7.1,25.8 -7.4,45.6 -9.6,55.9 -3.4,10.2 -0.7,8.7 -13.8,14.3 -11.9,4.3 -63.2,7.9 -111.7,7.6 -48.5,0.3 -99.8,-3.3 -111.6,-7.6 -13,-5.6 -10.3,-4.2 -13.8,-14.3 -2.3,-10.3 -2.5,-30.1 -9.6,-55.9 -10.6,-33.2 -45.9,-63.7 -78.7,-110.3 -32.9,-45.8 -62.9,-102.3 -62.9,-173.9 0,-71.4 31.9,-136.3 82.3,-183.7 50.4,-47.4 119,-76.6 190.9,-76.6h6.6c72,0 140.7,29.2 191,76.6 50.4,47.4 82.3,112.2 82.3,183.7 0,71.5 -29.9,128.1 -62.9,173.9zM381.2,824.7h261.7v-35.5L381.2,789.2v35.5zM381.2,891.8h261.7v-35.6L381.2,856.2v35.6zM454.6,958.9L569.5,958.9v-35.6L454.6,923.3v35.6z"/>
 </vector>
