+++ app/src/main/res/drawable/ic_lamp_open.xml
@@ -1,4 +1,4 @@
 <vector android:height="24dp" android:viewportHeight="1024.0"
     android:viewportWidth="1024.0" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FF000000" android:pathData="M507,89.7L507,62.3v27.4zM218.5,185.5l19.4,19.4 -19.4,-19.4zM101.7,476.8h27.4,-27.4zM225,765.3l19.4,-19.4 -19.4,19.4zM516.3,882.1v-27.4,27.4zM804.7,758.8L785.3,739.4l19.4,19.4zM921.5,467.6L894.2,467.6h27.4zM798.2,179l-19.4,19.4 19.4,-19.4zM511.6,156.9c-170.9,0 -315.4,150 -315.4,327.4 0,116.3 63.8,223.9 164.1,279.3v26.6c0,33.3 28.1,60.3 62.7,60.3h173.5c9.8,0 19.4,-2.1 26.1,-5.3 23.8,-9.3 39,-30.8 39,-55v-26.1c100.9,-55.3 165.1,-163 165.1,-280 0.1,-177.3 -144.4,-327.2 -315.2,-327.2m105.3,559.3l-18.1,8.5v62.4c-0.1,0 -0.4,0 -0.6,0.1 -0.5,0.1 -1,0.3 -1.6,0.3h-170.1l-2.6,2.8 -0.4,-65.9 -18.1,-8.5c-88.7,-41.7 -146.1,-132.6 -146.1,-231.5 0,-143.2 115.5,-264.3 252.2,-264.3s252.3,121.1 252.3,264.3c0.3,99.3 -57.4,190.4 -146.7,231.9M511.6,93.8c17.4,0 31.5,-14.1 31.5,-31.5L543.1,30.8C542.8,13.6 528.8,0 511.6,0S480.5,13.6 480.1,30.8v31.5c0,17.4 14.1,31.5 31.5,31.5m-289.8,119.8c12.8,0 24.3,-7.6 29.1,-19.5 4.9,-11.8 2.3,-25.4 -6.8,-34.4L221.8,137.6c-12.3,-12.3 -32.3,-12.3 -44.7,0 -12.3,12.3 -12.3,32.3 0,44.7l22.3,22.3c6.4,6.1 14.4,9.1 22.4,9.1zM101.7,440.3L70.2,440.3C53.2,440.9 39.8,454.8 39.8,471.8s13.4,30.9 30.4,31.5h31.5c17,-0.6 30.4,-14.5 30.4,-31.5S118.7,440.9 101.7,440.3m851,0.4h-31.5c-17.4,0 -31.5,14.1 -31.5,31.5 0,17.4 14.1,31.5 31.5,31.5h31.5c17.4,0 31.5,-14.1 31.5,-31.5 0,-17.5 -14.1,-31.5 -31.5,-31.5m-151.4,-302.8L779.1,160.1c-12.4,12.3 -12.4,32.3 0,44.7 12.3,12.4 32.3,12.4 44.7,0l22.3,-22.3c12.4,-12.3 12.4,-32.3 0,-44.7 -12.4,-12.4 -32.4,-12.4 -44.7,0M573.2,882h-122.7c-18.4,0 -33.3,14.9 -33.3,33.4 0,18.4 14.9,32.3 33.3,32.3h122.7c18.4,0 33.4,-13.9 33.4,-32.3 -0.1,-18.4 -15.1,-33.4 -33.4,-33.4m-20.6,94.7h-81.8c-12.3,0 -22.1,10.8 -22.1,24s9.9,23.3 22.1,23.3h81.8c12.3,0 22.1,-10 22.1,-23.3 0,-13.3 -9.9,-24 -22.1,-24"/>
+    <path android:fillColor="#ff669900" android:pathData="M507,89.7L507,62.3v27.4zM218.5,185.5l19.4,19.4 -19.4,-19.4zM101.7,476.8h27.4,-27.4zM225,765.3l19.4,-19.4 -19.4,19.4zM516.3,882.1v-27.4,27.4zM804.7,758.8L785.3,739.4l19.4,19.4zM921.5,467.6L894.2,467.6h27.4zM798.2,179l-19.4,19.4 19.4,-19.4zM511.6,156.9c-170.9,0 -315.4,150 -315.4,327.4 0,116.3 63.8,223.9 164.1,279.3v26.6c0,33.3 28.1,60.3 62.7,60.3h173.5c9.8,0 19.4,-2.1 26.1,-5.3 23.8,-9.3 39,-30.8 39,-55v-26.1c100.9,-55.3 165.1,-163 165.1,-280 0.1,-177.3 -144.4,-327.2 -315.2,-327.2m105.3,559.3l-18.1,8.5v62.4c-0.1,0 -0.4,0 -0.6,0.1 -0.5,0.1 -1,0.3 -1.6,0.3h-170.1l-2.6,2.8 -0.4,-65.9 -18.1,-8.5c-88.7,-41.7 -146.1,-132.6 -146.1,-231.5 0,-143.2 115.5,-264.3 252.2,-264.3s252.3,121.1 252.3,264.3c0.3,99.3 -57.4,190.4 -146.7,231.9M511.6,93.8c17.4,0 31.5,-14.1 31.5,-31.5L543.1,30.8C542.8,13.6 528.8,0 511.6,0S480.5,13.6 480.1,30.8v31.5c0,17.4 14.1,31.5 31.5,31.5m-289.8,119.8c12.8,0 24.3,-7.6 29.1,-19.5 4.9,-11.8 2.3,-25.4 -6.8,-34.4L221.8,137.6c-12.3,-12.3 -32.3,-12.3 -44.7,0 -12.3,12.3 -12.3,32.3 0,44.7l22.3,22.3c6.4,6.1 14.4,9.1 22.4,9.1zM101.7,440.3L70.2,440.3C53.2,440.9 39.8,454.8 39.8,471.8s13.4,30.9 30.4,31.5h31.5c17,-0.6 30.4,-14.5 30.4,-31.5S118.7,440.9 101.7,440.3m851,0.4h-31.5c-17.4,0 -31.5,14.1 -31.5,31.5 0,17.4 14.1,31.5 31.5,31.5h31.5c17.4,0 31.5,-14.1 31.5,-31.5 0,-17.5 -14.1,-31.5 -31.5,-31.5m-151.4,-302.8L779.1,160.1c-12.4,12.3 -12.4,32.3 0,44.7 12.3,12.4 32.3,12.4 44.7,0l22.3,-22.3c12.4,-12.3 12.4,-32.3 0,-44.7 -12.4,-12.4 -32.4,-12.4 -44.7,0M573.2,882h-122.7c-18.4,0 -33.3,14.9 -33.3,33.4 0,18.4 14.9,32.3 33.3,32.3h122.7c18.4,0 33.4,-13.9 33.4,-32.3 -0.1,-18.4 -15.1,-33.4 -33.4,-33.4m-20.6,94.7h-81.8c-12.3,0 -22.1,10.8 -22.1,24s9.9,23.3 22.1,23.3h81.8c12.3,0 22.1,-10 22.1,-23.3 0,-13.3 -9.9,-24 -22.1,-24"/>
 </vector>
