+++ app/src/main/res/drawable/ic_weather_rain.xml
@@ -1,6 +1,6 @@
 <vector android:height="24dp" android:viewportHeight="1024.0"
     android:viewportWidth="1024.0" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FF000000" android:pathData="M339.5,782.2l-0,0c0,0 -58,61 -63.2,110.3 -0,0.3 -0.1,0.6 -0.1,0.9 -0.1,0.7 -0.1,1.4 -0.2,2.1 -0,0.3 -0,0.6 -0.1,0.9 -0,1 -0.1,2 -0.1,3 0,36.1 29.7,65.3 66.4,65.3 36.6,0 66.4,-29.2 66.4,-65.3 0,-2.6 -0.2,-5.1 -0.5,-7.6C407.2,875.3 398.3,832.9 339.5,782.2z"/>
-    <path android:fillColor="#FF000000" android:pathData="M613.1,780.3l-0,0c0,0 -58.3,57.6 -63.5,104 -0,0.3 -0.1,0.6 -0.1,0.8 -0.1,0.7 -0.1,1.3 -0.2,2 -0,0.3 -0,0.6 -0.1,0.9 -0,0.9 -0.1,1.9 -0.1,2.8 0,34 29.8,61.6 66.6,61.6s66.6,-27.6 66.6,-61.6c0,-2.4 -0.2,-4.8 -0.5,-7.1C681,868 672,828.1 613.1,780.3z"/>
-    <path android:fillColor="#FF000000" android:pathData="M758.8,207.1c-49.4,-92 -146.5,-154.6 -258.3,-154.6 -120.5,0 -224,72.8 -269,176.8C131.3,250.6 56.1,339.5 56.1,446c0,122.3 99.2,221.5 221.5,221.5 2.8,0 5.6,-0.1 8.4,-0.2l58,0 0,22.7 0,0c-0,0.2 -0,0.3 -0,0.5 0,33 26.8,59.8 59.8,59.8 33,0 59.8,-26.8 59.8,-59.8 0,-0.2 -0,-0.3 -0,-0.5l0,0L463.6,667.3l136.4,0c-0.3,2.3 -0.4,4.5 -0.4,6.9 0,33 26.8,59.8 59.8,59.8 33,0 59.8,-26.8 59.8,-59.8 0,-2.3 -0.1,-4.6 -0.4,-6.9l35.9,0 0,-0.3c124.6,-2.8 224.8,-104.7 224.8,-230C979.4,313.1 881.4,212.1 758.8,207.1zM754.6,599.8c0,0 -16.8,0.4 -21,-0.1L277.2,599.7c-0.1,0 -0.2,0 -0.3,0 -0.1,0 -0.2,-0 -0.3,-0l-1.3,0 0,-0c-83.6,-0.9 -151.2,-68.9 -151.2,-152.7 0,-84.4 68.4,-152.8 152.8,-152.8 0.9,0 1.9,0 2.8,0 5.6,-24.3 15.1,-47.1 27.8,-67.7 -0,-0 -0.1,-0 -0.1,-0 13.2,-21.3 29.8,-40.4 49.1,-56.2 0.1,-0.1 0.2,-0.2 0.3,-0.3 1,-0.8 2.1,-1.7 3.1,-2.5 0.5,-0.4 1.1,-0.8 1.6,-1.3 0.5,-0.4 0.9,-0.7 1.4,-1.1 38.2,-29.3 86.1,-46.8 138,-46.8 101.5,0 187.5,66.7 216.4,158.6 11.3,-2.5 23,-3.8 35.1,-3.8 11.3,0 22.3,1.1 32.9,3.3 -0,-0 -0,-0 -0,-0 3.9,0.8 7.8,1.8 11.6,2.8 0.2,0.1 0.5,0.1 0.7,0.2 0.9,0.3 1.8,0.5 2.7,0.8 0.5,0.1 0.9,0.3 1.4,0.4 0.6,0.2 1.2,0.4 1.8,0.6 0.7,0.2 1.4,0.4 2,0.7 0.2,0.1 0.5,0.2 0.7,0.2 64,22.3 109.9,83.1 109.9,154.7C916.1,526.7 844,598.6 754.6,599.8z"/>
+    <path android:fillColor="#ccc" android:pathData="M339.5,782.2l-0,0c0,0 -58,61 -63.2,110.3 -0,0.3 -0.1,0.6 -0.1,0.9 -0.1,0.7 -0.1,1.4 -0.2,2.1 -0,0.3 -0,0.6 -0.1,0.9 -0,1 -0.1,2 -0.1,3 0,36.1 29.7,65.3 66.4,65.3 36.6,0 66.4,-29.2 66.4,-65.3 0,-2.6 -0.2,-5.1 -0.5,-7.6C407.2,875.3 398.3,832.9 339.5,782.2z"/>
+    <path android:fillColor="#ccc" android:pathData="M613.1,780.3l-0,0c0,0 -58.3,57.6 -63.5,104 -0,0.3 -0.1,0.6 -0.1,0.8 -0.1,0.7 -0.1,1.3 -0.2,2 -0,0.3 -0,0.6 -0.1,0.9 -0,0.9 -0.1,1.9 -0.1,2.8 0,34 29.8,61.6 66.6,61.6s66.6,-27.6 66.6,-61.6c0,-2.4 -0.2,-4.8 -0.5,-7.1C681,868 672,828.1 613.1,780.3z"/>
+    <path android:fillColor="#ccc" android:pathData="M758.8,207.1c-49.4,-92 -146.5,-154.6 -258.3,-154.6 -120.5,0 -224,72.8 -269,176.8C131.3,250.6 56.1,339.5 56.1,446c0,122.3 99.2,221.5 221.5,221.5 2.8,0 5.6,-0.1 8.4,-0.2l58,0 0,22.7 0,0c-0,0.2 -0,0.3 -0,0.5 0,33 26.8,59.8 59.8,59.8 33,0 59.8,-26.8 59.8,-59.8 0,-0.2 -0,-0.3 -0,-0.5l0,0L463.6,667.3l136.4,0c-0.3,2.3 -0.4,4.5 -0.4,6.9 0,33 26.8,59.8 59.8,59.8 33,0 59.8,-26.8 59.8,-59.8 0,-2.3 -0.1,-4.6 -0.4,-6.9l35.9,0 0,-0.3c124.6,-2.8 224.8,-104.7 224.8,-230C979.4,313.1 881.4,212.1 758.8,207.1zM754.6,599.8c0,0 -16.8,0.4 -21,-0.1L277.2,599.7c-0.1,0 -0.2,0 -0.3,0 -0.1,0 -0.2,-0 -0.3,-0l-1.3,0 0,-0c-83.6,-0.9 -151.2,-68.9 -151.2,-152.7 0,-84.4 68.4,-152.8 152.8,-152.8 0.9,0 1.9,0 2.8,0 5.6,-24.3 15.1,-47.1 27.8,-67.7 -0,-0 -0.1,-0 -0.1,-0 13.2,-21.3 29.8,-40.4 49.1,-56.2 0.1,-0.1 0.2,-0.2 0.3,-0.3 1,-0.8 2.1,-1.7 3.1,-2.5 0.5,-0.4 1.1,-0.8 1.6,-1.3 0.5,-0.4 0.9,-0.7 1.4,-1.1 38.2,-29.3 86.1,-46.8 138,-46.8 101.5,0 187.5,66.7 216.4,158.6 11.3,-2.5 23,-3.8 35.1,-3.8 11.3,0 22.3,1.1 32.9,3.3 -0,-0 -0,-0 -0,-0 3.9,0.8 7.8,1.8 11.6,2.8 0.2,0.1 0.5,0.1 0.7,0.2 0.9,0.3 1.8,0.5 2.7,0.8 0.5,0.1 0.9,0.3 1.4,0.4 0.6,0.2 1.2,0.4 1.8,0.6 0.7,0.2 1.4,0.4 2,0.7 0.2,0.1 0.5,0.2 0.7,0.2 64,22.3 109.9,83.1 109.9,154.7C916.1,526.7 844,598.6 754.6,599.8z"/>
 </vector>
