+++ app/src/main/res/drawable/ic_weather_sunny.xml
@@ -1,12 +1,12 @@
 <vector android:height="24dp" android:viewportHeight="1024.0"
     android:viewportWidth="1024.0" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FF000000" android:pathData="M688.7,335.3c-47.2,-47.2 -110,-73.2 -176.7,-73.2 -66.8,0 -129.5,26 -176.7,73.2C288.1,382.5 262.1,445.2 262.1,512c0,66.8 26,129.5 73.2,176.7 47.2,47.2 110,73.2 176.7,73.2 66.8,0 129.5,-26 176.7,-73.2 47.2,-47.2 73.2,-110 73.2,-176.7C761.9,445.2 735.9,382.5 688.7,335.3zM512,707.6c-107.9,0 -195.6,-87.8 -195.6,-195.6 0,-107.9 87.8,-195.6 195.6,-195.6S707.6,404.1 707.6,512C707.6,619.9 619.9,707.6 512,707.6z"/>
-    <path android:fillColor="#FF000000" android:pathData="M512,199.1c15,0 27.1,-12.1 27.1,-27.1L539.1,76.8c0,-15 -12.1,-27.1 -27.1,-27.1 -15,0 -27.1,12.1 -27.1,27.1l0,95.1C484.9,186.9 497,199.1 512,199.1z"/>
-    <path android:fillColor="#FF000000" android:pathData="M252.4,290.7c5.3,5.3 12.2,7.9 19.2,7.9 6.9,0 13.9,-2.7 19.2,-7.9 10.6,-10.6 10.6,-27.8 -0,-38.4l-67.3,-67.3c-10.6,-10.6 -27.8,-10.6 -38.4,0 -10.6,10.6 -10.6,27.8 0,38.4L252.4,290.7z"/>
-    <path android:fillColor="#FF000000" android:pathData="M199.1,512c0,-15 -12.1,-27.1 -27.1,-27.1L76.8,484.9c-15,0 -27.1,12.1 -27.1,27.1s12.1,27.1 27.1,27.1l95.1,0C186.9,539.1 199.1,527 199.1,512z"/>
-    <path android:fillColor="#FF000000" android:pathData="M252.3,733.3l-67.3,67.3c-10.6,10.6 -10.6,27.8 0,38.4 5.3,5.3 12.2,7.9 19.2,7.9 6.9,0 13.9,-2.7 19.2,-8l67.3,-67.3c10.6,-10.6 10.6,-27.8 -0,-38.4S262.9,722.7 252.3,733.3z"/>
-    <path android:fillColor="#FF000000" android:pathData="M512,824.9c-15,0 -27.1,12.1 -27.1,27.1L484.9,947.2c0,15 12.1,27.1 27.1,27.1 15,0 27.1,-12.1 27.1,-27.1l0,-95.1C539.1,837.1 527,824.9 512,824.9z"/>
-    <path android:fillColor="#FF000000" android:pathData="M771.7,733.3c-10.6,-10.6 -27.8,-10.6 -38.4,0 -10.6,10.6 -10.6,27.8 0,38.4l67.3,67.3c5.3,5.3 12.2,7.9 19.2,7.9 6.9,0 13.9,-2.7 19.2,-8 10.6,-10.6 10.6,-27.8 -0,-38.4L771.7,733.3z"/>
-    <path android:fillColor="#FF000000" android:pathData="M947.2,484.9l-95.1,0c-15,0 -27.1,12.1 -27.1,27.1s12.1,27.1 27.1,27.1L947.2,539.1c15,0 27.1,-12.1 27.1,-27.1S962.2,484.9 947.2,484.9z"/>
-    <path android:fillColor="#FF000000" android:pathData="M752.5,298.7c6.9,0 13.9,-2.7 19.2,-8l67.3,-67.3c10.6,-10.6 10.6,-27.8 -0,-38.4 -10.6,-10.6 -27.8,-10.6 -38.4,0l-67.3,67.3c-10.6,10.6 -10.6,27.8 0,38.4C738.6,296 745.5,298.7 752.5,298.7z"/>
+    <path android:fillColor="#9C27B0" android:pathData="M688.7,335.3c-47.2,-47.2 -110,-73.2 -176.7,-73.2 -66.8,0 -129.5,26 -176.7,73.2C288.1,382.5 262.1,445.2 262.1,512c0,66.8 26,129.5 73.2,176.7 47.2,47.2 110,73.2 176.7,73.2 66.8,0 129.5,-26 176.7,-73.2 47.2,-47.2 73.2,-110 73.2,-176.7C761.9,445.2 735.9,382.5 688.7,335.3zM512,707.6c-107.9,0 -195.6,-87.8 -195.6,-195.6 0,-107.9 87.8,-195.6 195.6,-195.6S707.6,404.1 707.6,512C707.6,619.9 619.9,707.6 512,707.6z"/>
+    <path android:fillColor="#9C27B0" android:pathData="M512,199.1c15,0 27.1,-12.1 27.1,-27.1L539.1,76.8c0,-15 -12.1,-27.1 -27.1,-27.1 -15,0 -27.1,12.1 -27.1,27.1l0,95.1C484.9,186.9 497,199.1 512,199.1z"/>
+    <path android:fillColor="#9C27B0" android:pathData="M252.4,290.7c5.3,5.3 12.2,7.9 19.2,7.9 6.9,0 13.9,-2.7 19.2,-7.9 10.6,-10.6 10.6,-27.8 -0,-38.4l-67.3,-67.3c-10.6,-10.6 -27.8,-10.6 -38.4,0 -10.6,10.6 -10.6,27.8 0,38.4L252.4,290.7z"/>
+    <path android:fillColor="#9C27B0" android:pathData="M199.1,512c0,-15 -12.1,-27.1 -27.1,-27.1L76.8,484.9c-15,0 -27.1,12.1 -27.1,27.1s12.1,27.1 27.1,27.1l95.1,0C186.9,539.1 199.1,527 199.1,512z"/>
+    <path android:fillColor="#9C27B0" android:pathData="M252.3,733.3l-67.3,67.3c-10.6,10.6 -10.6,27.8 0,38.4 5.3,5.3 12.2,7.9 19.2,7.9 6.9,0 13.9,-2.7 19.2,-8l67.3,-67.3c10.6,-10.6 10.6,-27.8 -0,-38.4S262.9,722.7 252.3,733.3z"/>
+    <path android:fillColor="#9C27B0" android:pathData="M512,824.9c-15,0 -27.1,12.1 -27.1,27.1L484.9,947.2c0,15 12.1,27.1 27.1,27.1 15,0 27.1,-12.1 27.1,-27.1l0,-95.1C539.1,837.1 527,824.9 512,824.9z"/>
+    <path android:fillColor="#9C27B0" android:pathData="M771.7,733.3c-10.6,-10.6 -27.8,-10.6 -38.4,0 -10.6,10.6 -10.6,27.8 0,38.4l67.3,67.3c5.3,5.3 12.2,7.9 19.2,7.9 6.9,0 13.9,-2.7 19.2,-8 10.6,-10.6 10.6,-27.8 -0,-38.4L771.7,733.3z"/>
+    <path android:fillColor="#9C27B0" android:pathData="M947.2,484.9l-95.1,0c-15,0 -27.1,12.1 -27.1,27.1s12.1,27.1 27.1,27.1L947.2,539.1c15,0 27.1,-12.1 27.1,-27.1S962.2,484.9 947.2,484.9z"/>
+    <path android:fillColor="#9C27B0" android:pathData="M752.5,298.7c6.9,0 13.9,-2.7 19.2,-8l67.3,-67.3c10.6,-10.6 10.6,-27.8 -0,-38.4 -10.6,-10.6 -27.8,-10.6 -38.4,0l-67.3,67.3c-10.6,10.6 -10.6,27.8 0,38.4C738.6,296 745.5,298.7 752.5,298.7z"/>
 </vector>
